Getting rid of process motion return type.
