+++ rules.mk
@@ -1,5 +1,6 @@
 # note that the order is important here
 SRC += qmk-vim/src/process_func.c
+SRC += qmk-vim/src/numbered_actions.c
 SRC += qmk-vim/src/motions.c
 SRC += qmk-vim/src/actions.c
 SRC += qmk-vim/src/modes.c
